project6/requirements.txt
asgiref==3.8.1
django==5.2.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
gunicorn==23.0.0
sqlparse==0.5.3
tzdata==2025.2
